taking out the script check from module and add it to main es search
