+++ lib/scripts/sampler.js
@@ -4,31 +4,18 @@ var _ = require('lodash');
 
 var Sampler = function () {};
 
-Sampler.prototype.checkAndSample = function(es_url, index, type, esQuery, aggregationObjects, query) {
-    var params=[];
-    
-    query['include'] && params.push("include="+ query['include']);
-    query['limit'] && params.push("size="+ query['limit']);
-    query['offset'] && params.push("from="+ query['offset']);
-
-    var queryStr = '?'+params.join('&');
-
-    var es_resource = es_url + '/' + index + '/' + type + '/_search' + queryStr;
-
+Sampler.prototype.sample = function(index, type, esQuery, aggregationObjects, query, es_resource) {
     return Promise.resolve()
     .then(function() {
         // if sampleAggregation make the query to see how many hits we have
-        if(query.script === 'sampler') {
-            var countUrl = es_url + '/' + index + '/' + type + '/_count' + queryStr;
-            var countQuery = _.clone(JSON.parse(esQuery));
-        
-            delete countQuery.aggs;
-            return $http({url : countUrl, method: 'GET', body: JSON.stringify(countQuery)});
-        }
+        var countUrl = es_resource.replace('/_search/', '/_count/');
+        var countQuery = _.clone(JSON.parse(esQuery));
+    
+        delete countQuery.aggs;
+        return $http({url : countUrl, method: 'GET', body: JSON.stringify(countQuery)});
 
     })
     .spread(function(response) {
-        if (query.script !== 'sampler') return $http({url : es_resource, method: 'GET', body: esQuery});
         var es_results;
         response && response.body && (es_results = JSON.parse(response.body));
         var total = es_results && es_results.count;
